Modify calibration before FP: sending N, facing down before and up after
Send command to let the robot face down after returning to start zone when exploration is done. Send N to arduino to do calibration and after the calibration, the robot will face up and ready for fastest path. This is for the special calibration before the fastest path just in case the robot does not return to the centre of 3x3 but the border, so special align_front (with long distance) will attract the robot to the wall.